+++ templates/home/index.tsx
@@ -1,5 +1,5 @@
 import { Divider } from "@chakra-ui/react";
-import { MetaTitle } from "@common/meta";
+import { MetaTitle } from "@common/components";
 import { About } from "./About";
 import { Collage } from "./Collage";
 import { Hero } from "./Hero";
